+++ packages/edit/i18n/en.js
@@ -10,6 +10,7 @@ i18n.map('en', {
 	btn_reload: "Reload",
 
 	label_editplaces: "Edits",
+	label_owner: "Owner",
 
 	edit_error_notOwner: "Not Owner",
 	edit_error_notUrl: "Url Not Valid"