+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - plugin geoinfo show country and city in place search results 
 - core fix zoom min for cluster disable
 - core new Kepler module K.Alert.xxx to show notifications over the map(extend sAlert lib)
+- plugin cats moved typeahead.js in kepler plugin edit
 
 ## v1.6.5
 - new tab in panel place show nearby places